requirements.txt
requests
pandas
SQLAlchemy
streamlit
python-dotenv     
